Move get_worker_thread_num definition to thread_num_tss files
