+++ libs/threading_base/include/hpx/threading_base/thread_num_tss.hpp
@@ -8,6 +8,7 @@
 #define HPX_RUNTIME_THREADS_DETAIL_THREAD_NUM_TSS_JUL_17_2015_0811PM
 
 #include <hpx/config.hpp>
+#include <hpx/errors.hpp>
 
 #include <cstddef>
 #include <cstdint>
@@ -53,6 +54,43 @@ namespace hpx { namespace threads { namespace detail {
         std::size_t thread_num_;
     };
 }}}    // namespace hpx::threads::detail
+
+namespace hpx
+{
+    ///////////////////////////////////////////////////////////////////////////
+    /// \brief Return the number of the current OS-thread running in the
+    ///        runtime instance the current HPX-thread is executed with.
+    ///
+    /// This function returns the zero based index of the OS-thread which
+    /// executes the current HPX-thread.
+    ///
+    /// \note   The returned value is zero based and its maximum value is
+    ///         smaller than the overall number of OS-threads executed (as
+    ///         returned by \a get_os_thread_count().
+    ///
+    /// \note   This function needs to be executed on a HPX-thread. It will
+    ///         fail otherwise (it will return -1).
+    HPX_API_EXPORT std::size_t get_worker_thread_num();
+
+    ///////////////////////////////////////////////////////////////////////////
+    /// \brief Return the number of the current OS-thread running in the
+    ///        runtime instance the current HPX-thread is executed with.
+    ///
+    /// This function returns the zero based index of the OS-thread which
+    /// executes the current HPX-thread.
+    ///
+    /// \param ec [in,out] this represents the error status on exit.
+    ///
+    /// \note   The returned value is zero based and its maximum value is
+    ///         smaller than the overall number of OS-threads executed (as
+    ///         returned by \a get_os_thread_count(). It will return -1 if
+    ///         the current thread is not a known thread or if the runtime is
+    ///         not in running state.
+    ///
+    /// \note   This function needs to be executed on a HPX-thread. It will
+    ///         fail otherwise (it will return -1).
+    HPX_API_EXPORT std::size_t get_worker_thread_num(error_code& ec);
+}
 
 #include <hpx/config/warnings_suffix.hpp>
